--- WeatherApi-XUTest/City.cs
@@ -1,5 +1,0 @@
-﻿public class City
-{
-    public int Id { get; set; }
-    public string Name { get; set; }
-}
