--- WeatherApi-XUTest/DataContext.cs
@@ -1,9 +1,0 @@
-﻿using Microsoft.EntityFrameworkCore;
-
-public class DataContext : DbContext
-{
-    public DataContext(DbContextOptions options) : base(options)
-    {
-    }
-    public DbSet<City> Cities { get; set; }
-}
