+++ WeatherApi_CI-CD/Data/DataContext.cs
@@ -1,8 +1,9 @@
 ﻿using Microsoft.EntityFrameworkCore;
+using WeatherApi_CI_CD.Model;
 
-namespace WeatherApi_CI_CD
+namespace WeatherApi_CI_CD.Data
 {
-    public class DataContext:DbContext
+    public class DataContext : DbContext
     {
         public DataContext(DbContextOptions<DataContext> options) : base(options)
         {
@@ -11,3 +12,5 @@ namespace WeatherApi_CI_CD
         public DbSet<City> Cities { get; set; }
     }
 }
+
+
